requirements.txt
asgiref==3.2.3
cffi==1.14.0
cryptography==2.9.2
Django==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
numpy==1.19.0
pandas==1.0.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.0
pillow==7.2.0
zappa==0.51.0
sendgrid==6.4.4
django-cors-headers==3.4.0
twilio==6.46.0